Use fabs to fix Xcode 8.3 warnings

+++ Source/AMPopTip.m
@@ -162,9 +162,9 @@
         
         if (self.bubbleOffset > 0 && arrowPosition.x < self.bubbleOffset) {
             self.bubbleOffset = arrowPosition.x - self.arrowSize.width;
-        } else if (self.bubbleOffset < 0 && frame.size.width < fabsf(self.bubbleOffset)) {
+        } else if (self.bubbleOffset < 0 && frame.size.width < fabs(self.bubbleOffset)) {
             self.bubbleOffset = -(arrowPosition.x - self.arrowSize.width);
-        } else if (self.bubbleOffset < 0 && (frame.origin.x - arrowPosition.x) < fabsf(self.bubbleOffset)) {
+        } else if (self.bubbleOffset < 0 && (frame.origin.x - arrowPosition.x) < fabs(self.bubbleOffset)) {
             self.bubbleOffset = -(self.arrowSize.width + self.edgeMargin);
         }
         
@@ -172,7 +172,7 @@
         CGFloat leftSpace = frame.origin.x - self.containerView.frame.origin.x;
         CGFloat rightSpace = self.containerView.frame.size.width - leftSpace - frame.size.width;
         
-        if (self.bubbleOffset < 0 && leftSpace < fabsf(self.bubbleOffset)) {
+        if (self.bubbleOffset < 0 && leftSpace < fabs(self.bubbleOffset)) {
             self.bubbleOffset = -leftSpace + self.edgeMargin;
         } else if (self.bubbleOffset > 0 && rightSpace < self.bubbleOffset) {
             self.bubbleOffset = rightSpace - self.edgeMargin;
@@ -206,7 +206,7 @@
         
         if (self.bubbleOffset > 0 && arrowPosition.y < self.bubbleOffset) {
             self.bubbleOffset = arrowPosition.y - self.arrowSize.width;
-        } else if (self.bubbleOffset < 0 && frame.size.height < fabsf(self.bubbleOffset)) {
+        } else if (self.bubbleOffset < 0 && frame.size.height < fabs(self.bubbleOffset)) {
             self.bubbleOffset = -(arrowPosition.y - self.arrowSize.height);
         }
         
@@ -214,7 +214,7 @@
         CGFloat topSpace = frame.origin.y - self.containerView.frame.origin.y;
         CGFloat bottomSpace = self.containerView.frame.size.height - topSpace - frame.size.height;
         
-        if (self.bubbleOffset < 0 && topSpace < fabsf(self.bubbleOffset)) {
+        if (self.bubbleOffset < 0 && topSpace < fabs(self.bubbleOffset)) {
             self.bubbleOffset = -topSpace + self.edgeMargin;
         } else if (self.bubbleOffset > 0 && bottomSpace < self.bubbleOffset) {
             self.bubbleOffset = bottomSpace - self.edgeMargin;
